+++ CHANGELOG.md
@@ -93,3 +93,18 @@ removing accounts, adding and removing money, calculating interest, getting
 the type of an account, getting values of accounts, adding a overdraft fee,
 getting the total number of accounts, and getting the total money in all
 accounts.
+
+## [2024.10.17]
+
+### Added
+
+- Designed kernel and enhanced interfaces for BankAccount component
+
+### Updated
+
+- Changed design to include all method contracts for the implementations of all
+the method from the proof of concept
+- Extended Standard into BankAccountKernel and included 4 methods in the
+kernel
+- Extended BankAccountKernel into BankAccount and included 6 more methods in
+the BankAccount
